Change invoice item quantity data type int to decimal

+++ LBH.AdultSocialCare.Transactions.Api/V1/Boundary/InvoiceBoundaries/Request/InvoiceItemForCreationRequest.cs
@@ -7,7 +7,7 @@ namespace LBH.AdultSocialCare.Transactions.Api.V1.Boundary.InvoiceBoundaries.Req
     {
         [Required] public string ItemName { get; set; }
         [Required] public decimal? PricePerUnit { get; set; }
-        [Required] public int? Quantity { get; set; }
+        [Required] public decimal? Quantity { get; set; }
         public Guid? SupplierReturnItemId { get; set; }
         [Required] public Guid? CreatorId { get; set; }
     }
